Move app to components, url container to containers
