+++ src/apiCalls.js
@@ -1,6 +1,6 @@
 export const getUrls = () => {
   return fetch('http://localhost:3001/api/v1/urls')
-      .then(response => response.json())
+    .then(response => response.json())
 }
 
 export const getShortUrl = (urlInfo) => {
